requirements.txt
djangorestframework
django<3.0
requests
discord.py
django-environ
gunicorn